refactored names for package and test package
