+++ molgenis-ontology/src/main/java/org/molgenis/ontology/tree/OntologyRepositoryRegistrator.java
@@ -1,4 +1,4 @@
-package org.molgenis.omx.biobankconnect.ontologytree;
+package org.molgenis.ontology.tree;
 
 import org.molgenis.data.DataService;
 import org.molgenis.data.elasticsearch.SearchService;
