+++ molgenis-ontology/src/test/java/org/molgenis/molgenis/utils/OntologyLoaderTest.java
@@ -1,4 +1,4 @@
-package org.molgenis.omx.biobankconnect.utils;
+package org.molgenis.molgenis.utils;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
